Extract common types to `metabase-types`
